Factor out the bottom-up postorder traversal logic.
